Disconnection warning from client to server

+++ shared/message.c
@@ -17,7 +17,7 @@ int receive_message(int sockFD, Message* msg)
     unsigned char buf1[5];
     ssize_t n_bytes = recv(sockFD, buf1, 5, 0);
     if(n_bytes <= 0)
-        return n_bytes;
+        return (int)n_bytes;
 
     msg->messageID = (int8_t)buf1[0];
 
@@ -33,7 +33,7 @@ int receive_message(int sockFD, Message* msg)
         n_bytes = recv(sockFD, (void*) msg->firstArg, firstArgSize * sizeof(char), 0);
         msg->firstArg[firstArgSize-1] = '\0';
         if(n_bytes <= 0)
-            return n_bytes;
+            return (int)n_bytes;
     }
 
     msg->secondArg = NULL;
@@ -43,7 +43,7 @@ int receive_message(int sockFD, Message* msg)
         n_bytes = recv(sockFD, (void*) msg->secondArg, secondArgSize * sizeof(char), 0);
         msg->secondArg[secondArgSize-1] = '\0';
         if(n_bytes <= 0)
-            return n_bytes;
+            return (int)n_bytes;
     }
 
     return 1;
@@ -63,18 +63,19 @@ int send_message(int sockFD, Message msg)
     //Send the message header
     ssize_t n_bytes = send(sockFD, buffer_msg_header, 5, 0);
     if(n_bytes <= 0)
-        return n_bytes;
+        return (int)n_bytes;
+        
     if(msg.firstArg != NULL)
     {
         n_bytes = send(sockFD, msg.firstArg, firstArgSize*sizeof(char), 0);
         if(n_bytes <= 0)
-            return n_bytes;
+            return (int)n_bytes;
     }
     if(msg.secondArg != NULL)
     {
         n_bytes = send(sockFD, msg.secondArg, secondArgSize*sizeof(char), 0);
         if(n_bytes <= 0)
-            return n_bytes;
+            return (int)n_bytes;
     }
     return 1;
 }
@@ -168,7 +169,7 @@ int send_auth_message(AuthMessage msg, int sock, struct sockaddr_in server_addre
     if(send_status != AUTH_MSG_BUFFER_SIZE)
     {
         //Didn't send correctly, send the send_status, except if 0>send_status>AUTH_MSG_BUFFER_SIZE, then send -1
-        return send_status > 0 ? -1 : send_status;
+        return send_status > 0 ? -1 : (int)send_status;
     }
 
     return 1;
